+++ user/models/Post.js
@@ -4,10 +4,11 @@ const PostSchema = mongoose.Schema({
     email: String,
     password: String,
     confirmPassword: String,
-    phoneNumber: String,
+    phoneNumber: Number,
     gender: String,
     firstName: String,
     lastName: String,
+    sessionToken: String,
     resetToken: String,
     resetTimeout:Date
 });
